# rewrite SizeOfValueInfo in Optimizer.fs to be tail-recursive