## Description

Pretty straight forward rewrite to be tail recursive and save some cycles.
The only way I see for `SizeOfValueInfo` to return something negative would be the
`| SizeValue (vdepth, _v) -> vdepth`
case, so I moved the `max` call there.
It's likely too small of an improvement to mention it in the release notes.

# This PR:
| Method    | Source            | Mean             | Error           | StdDev          | Gen0        | Gen1       | Gen2       | Allocated      |
|---------- |------------------ |-----------------:|----------------:|----------------:|------------:|-----------:|-----------:|---------------:|
| CompileCE | CE100xnest1.fs    |     595,608.2 us |     7,492.71 us |     7,008.69 us |   1000.0000 |          - |          - |   249126.77 KB |
| CompileCE | CE100xnest5.fs    |   6,112,593.8 us |    43,582.18 us |    38,634.46 us |  22000.0000 |  4000.0000 |  3000.0000 |  2208092.63 KB |
| CompileCE | CE200xnest5.fs    | 113,563,126.0 us | 2,052,977.45 us | 1,714,329.15 us | 328000.0000 | 97000.0000 | 63000.0000 | 12184176.05 KB |
| CompileCE | CEwCO100xnest5.fs |     385,267.2 us |     7,399.69 us |     6,921.68 us |           - |          - |          - |   272500.27 KB |
| CompileCE | CEwCO500xnest1.fs |   2,503,880.7 us |    18,760.53 us |    16,630.72 us |   2000.0000 |  1000.0000 |          - |   577406.52 KB |

# main:
| Method    | Source            | Mean             | Error         | StdDev        | Gen0        | Gen1       | Gen2       | Allocated      |
|---------- |------------------ |-----------------:|--------------:|--------------:|------------:|-----------:|-----------:|---------------:|
| CompileCE | CE100xnest1.fs    |     600,332.2 us |   4,924.26 us |   4,365.23 us |   2000.0000 |  1000.0000 |  1000.0000 |   245788.45 KB |
| CompileCE | CE100xnest5.fs    |   6,193,411.1 us |  24,108.82 us |  22,551.40 us |  19000.0000 |  4000.0000 |  2000.0000 |  2211443.17 KB |
| CompileCE | CE200xnest5.fs    | 115,246,799.8 us | 739,218.66 us | 655,298.08 us | 328000.0000 | 65000.0000 | 62000.0000 | 12184072.21 KB |
| CompileCE | CEwCO100xnest5.fs |     383,390.0 us |   3,427.27 us |   2,675.79 us |           - |          - |          - |   272443.69 KB |
| CompileCE | CEwCO500xnest1.fs |   2,509,841.0 us |  20,470.25 us |  19,147.88 us |   3000.0000 |  2000.0000 |  1000.0000 |   574606.42 KB |


## Checklist

- [ ] Test cases added
- [X] Performance benchmarks added in case of performance changes
- [ ] Release notes entry updated:
    > Please make sure to add an entry with short succinct description of the change as well as link to this pull request to the respective release notes file, if applicable.
    >
    > Release notes files:
    > - If anything under `src/Compiler` has been changed, please make sure to make an entry in `docs/release-notes/.FSharp.Compiler.Service/<version>.md`, where `<version>` is usually "highest" one, e.g. `42.8.200`
    > - If language feature was added (i.e. `LanguageFeatures.fsi` was changed), please add it to `docs/releae-notes/.Language/preview.md`
    > - If a change to `FSharp.Core` was made, please make sure to edit `docs/release-notes/.FSharp.Core/<version>.md` where version is "highest" one, e.g. `8.0.200`.

    > Information about the release notes entries format can be found in the [documentation](https://fsharp.github.io/fsharp-compiler-docs/release-notes/About.html).
    > Example:
    > * More inlines for Result module. ([PR #16106](https://github.com/dotnet/fsharp/pull/16106))
    > * Correctly handle assembly imports with public key token of 0 length. ([Issue #16359](https://github.com/dotnet/fsharp/issues/16359), [PR #16363](https://github.com/dotnet/fsharp/pull/16363))
    > *`while!` ([Language suggestion #1038](https://github.com/fsharp/fslang-suggestions/issues/1038), [PR #14238](https://github.com/dotnet/fsharp/pull/14238))

    > **If you believe that release notes are not necessary for this PR, please add `NO_RELEASE_NOTES` label to the pull request.**